modules/old/old_modules_v1/visualization/requirements.txt
streamlit
pandas
plotly
imbalanced-learn
numpy
pillow
requests
fastapi
uvicorn[standard]
pydantic
python-multipart
scikit-learn
redis
python-jose
aiofiles
psutil
typer